Move OrSuggestions to its own header
Prevents a recursive inclusion
