+++ src/libexpr/eval-cache.hh
@@ -3,6 +3,7 @@
 #include "sync.hh"
 #include "hash.hh"
 #include "eval.hh"
+#include "or-suggestions.hh"
 
 #include <functional>
 #include <variant>
